Add support for YMD75 rev 2 (#8853)
