+++ keyboards/ymd75/rev1/config.h
@@ -19,23 +19,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-#include "config_common.h"
-
-#define VENDOR_ID       0x20A0
-#define PRODUCT_ID      0x422D
-#define DEVICE_VER      0x0200
-
-#define MANUFACTURER    YMDK
-#define PRODUCT         YMD75 / MT84
-#define DESCRIPTION     75% Keyboard
-
-
-#define BACKLIGHT_PIN D4
-#define BACKLIGHT_LEVELS 12
-
-#define RGB_DI_PIN E2
-#define RGBLED_NUM 16
-#define RGBLIGHT_ANIMATIONS
-#define RGBLIGHT_HUE_STEP 12
-#define RGBLIGHT_SAT_STEP 15
-#define RGBLIGHT_VAL_STEP 18
+/* matrix size */
+#define MATRIX_ROWS 8
+#define MATRIX_COLS 15
+#define MATRIX_ROW_PINS { B0, B1, B2, B3, B4, B5, B6, B7 }
+#define MATRIX_COL_PINS { A0, A1, A2, A3, A4, A5, A6, A7, C7, C6, C5, C4, C3, C2, D7 }
+#define DIODE_DIRECTION COL2ROW
