+++ keyboards/ymd75/rev2/config.h
@@ -1,0 +1,8 @@
+#pragma once
+
+/* matrix size */
+#define MATRIX_ROWS 6
+#define MATRIX_COLS 15
+#define MATRIX_ROW_PINS { B7, B6, B5, B4, B3, B0 }
+#define MATRIX_COL_PINS { A0, A1, A2, A3, A4, A5, A6, A7, C7, C6, C5, C4, C3, C2, D7 }
+#define DIODE_DIRECTION COL2ROW
